requirements.txt
Flask
Pillow
pyproj
requests
Flask-Cors